Refactored parts of the new UI injection code
