+++ src/App.vue
@@ -133,7 +133,7 @@ textarea {
 		<connection-dialog></connection-dialog>
 		<messagebox-dialog></messagebox-dialog>
 
-		<component v-for="component in injectedComponents" :is="component" :key="component"></component>
+		<component v-for="component in injectedComponentNames" :is="component" :key="component"></component>
 	</v-app>
 </template>
 
@@ -143,7 +143,6 @@ textarea {
 import Piecon from 'piecon'
 import { mapState, mapGetters, mapActions } from 'vuex'
 
-import { setInjectComponent } from './store/uiinjection'
 import { Menu, Routes } from './routes'
 import { isPrinting } from './store/machine/modelEnums.js'
 
@@ -159,7 +158,9 @@ export default {
 			status: state => state.machine.model.state.status,
 
 			darkTheme: state => state.settings.darkTheme,
-			webcam: state => state.settings.webcam
+			webcam: state => state.settings.webcam,
+
+			injectedComponents: state => state.uiInjection.injectedComponents
 		}),
 		...mapGetters('machine', ['hasTemperaturesToDisplay']),
 		...mapGetters('machine/model', ['jobProgress']),
@@ -191,9 +192,9 @@ export default {
 	data() {
 		return {
 			drawer: this.$vuetify.breakpoint.lgAndUp,
-			injectedComponents: [],
 			hideGlobalContainer: false,
-			wasXs: this.$vuetify.breakpoint.xsOnly
+			wasXs: this.$vuetify.breakpoint.xsOnly,
+			injectedComponentNames: []
 		}
 	},
 	methods: {
@@ -215,10 +216,6 @@ export default {
 			if (document.title !== title) {
 				document.title = title;
 			}
-		},
-		injectComponent(name, component) {
-			this.$options.components[name] = component;
-			this.injectedComponents.push(name);
 		}
 	},
 	mounted() {
@@ -249,9 +246,6 @@ export default {
 			shadow: '#fff',			// Outer ring color
 			fallback: false			// Toggles displaying percentage in the title bar (possible values - true, false, 'force')
 		});
-
-		// Register function to inject custom components
-		setInjectComponent(this.injectComponent);
 	},
 	watch: {
 		currentPageCondition(to) {
@@ -282,6 +276,14 @@ export default {
 				Piecon.setProgress(to * 100);
 			}
 			this.updateTitle();
+		},
+		injectedComponents() {
+			this.injectedComponents.forEach(function(item) {
+				if (this.injectedComponentNames.indexOf(item.name) === -1) {
+					this.$options.components[item.name] = item.component;
+					this.injectedComponentNames.push(item.name);
+				}
+			}, this);
 		}
 	}
 }
