+++ src/store/uiInjection.js
@@ -1,0 +1,20 @@
+'use strict'
+
+export default  {
+	namespaced: true,
+	state: {
+		contextMenuItems: {
+			jobFileList: []
+		},
+		injectedComponents: []
+	},
+	mutations : {
+		registerPluginContextMenuItem(state, payload) {
+			state.contextMenuItems[payload.menu].push(payload.item);
+		},
+		injectComponent(state, payload) {
+			state.injectedComponents.push(payload);
+		}
+	}
+}
+
